review round: F1 read-only grant hole closed via TryCommitAtomic channel + F2 lock files + F3 de-commented durability grep + F4 no-throw refusals + F5 file-level key-set + F7 depth ceiling (A-C7-13) + F8/F9/F10/F11/F12

+++ unity/Assets/Scripts/Application/Save/SaveHeader.cs
@@ -78,6 +78,6 @@ namespace CatMetro.Application.Save
             (ushort)(b[at] | (b[at + 1] << 8));
 
         private static uint ReadU32(byte[] b, int at) =>
-            (uint)(b[at] | (b[at + 1] << 8) | (b[at + 2] << 16) | ((uint)b[at + 3] << 24));
+            (uint)b[at] | ((uint)b[at + 1] << 8) | ((uint)b[at + 2] << 16) | ((uint)b[at + 3] << 24);
     }
 }
